modelsをばらしてそれぞれのresolversの近くに
