+++ src/resolvers/NicovideoVideoSource/index.ts
@@ -2,9 +2,9 @@ import { GraphQLError } from "graphql";
 import { DataSource } from "typeorm";
 
 import { Video } from "../../db/entities/videos.js";
-import { VideoModel } from "../../graphql/models.js";
 import { Resolvers } from "../../graphql/resolvers.js";
 import { buildGqlId } from "../../utils/id.js";
+import { VideoModel } from "../video/model.js";
 
 export const resolveNicovideoVideoSource = ({
   dataSource,
