+++ src/resolvers/mutation/add_video_to_mylist.ts
@@ -6,10 +6,10 @@ import { ulid } from "ulid";
 import { MylistRegistration } from "../../db/entities/mylist_registrations.js";
 import { Mylist, MylistShareRange as MylistEntityShareRange } from "../../db/entities/mylists.js";
 import { Video } from "../../db/entities/videos.js";
-import { MylistRegistrationModel } from "../../graphql/models.js";
 import { MutationResolvers } from "../../graphql/resolvers.js";
 import { addVideoToMylist as addVideoToMylistInNeo4j } from "../../neo4j/add_video_to_mylist.js";
 import { ObjectType, removeIDPrefix } from "../../utils/id.js";
+import { MylistRegistrationModel } from "../mylist_registration/models.js";
 import { MYLIST_NOT_FOUND_OR_PRIVATE_ERROR, MYLIST_NOT_HOLDED_BY_YOU } from "../query/get_mylist.js";
 
 export const addVideoToMylist =
