+++ src/resolvers/mutation/register_videos.ts
@@ -9,10 +9,10 @@ import { VideoTag } from "../../db/entities/video_tags.js";
 import { VideoThumbnail } from "../../db/entities/video_thumbnails.js";
 import { VideoTitle } from "../../db/entities/video_titles.js";
 import { Video } from "../../db/entities/videos.js";
-import { VideoModel } from "../../graphql/models.js";
 import { MutationResolvers, RegisterVideoInputSourceType } from "../../graphql/resolvers.js";
 import { registerVideo as registerVideoInNeo4j } from "../../neo4j/register_video.js";
 import { ObjectType, removeIDPrefix } from "../../utils/id.js";
+import { VideoModel } from "../video/model.js";
 
 export const isValidNicovideoSourceId = (id: string): boolean => /[a-z]{2}\d+/.test(id);
