+++ src/resolvers/query/findNicovideoVideoSource.ts
@@ -1,7 +1,7 @@
 import { DataSource } from "typeorm";
 
 import { NicovideoVideoSource } from "../../db/entities/nicovideo_source.js";
-import { NicovideoVideoSourceModel } from "../../graphql/models.js";
+import { NicovideoVideoSourceModel } from "../NicovideoVideoSource/model.js";
 
 export const findNicovideoVideoSource =
   ({ dataSource: ds }: { dataSource: DataSource }) =>
