+++ src/resolvers/query/find_mylist.ts
@@ -1,9 +1,9 @@
 import { DataSource } from "typeorm";
 
 import { Mylist, MylistShareRange as MylistEntityShareRange } from "../../db/entities/mylists.js";
-import { MylistModel } from "../../graphql/models.js";
 import { QueryResolvers } from "../../graphql/resolvers.js";
 import { ObjectType, removeIDPrefix } from "../../utils/id.js";
+import { MylistModel } from "../mylist/model.js";
 
 export const MYLIST_NOT_FOUND_OR_PRIVATE_ERROR = "Mylist Not Found or Private";
 export const MYLIST_NOT_HOLDED_BY_YOU = "This mylist is not holded by you";
