+++ src/resolvers/query/get_user.ts
@@ -2,8 +2,8 @@ import { GraphQLError } from "graphql";
 import { DataSource } from "typeorm";
 
 import { User } from "../../db/entities/users.js";
-import { UserModel } from "../../graphql/models.js";
 import { QueryResolvers } from "../../graphql/resolvers.js";
+import { UserModel } from "../user/model.js";
 
 export const getUser =
   ({ dataSource }: { dataSource: DataSource }): QueryResolvers["user"] =>
