+++ src/resolvers/query/search_tags.ts
@@ -3,8 +3,8 @@ import { DataSource, In, Like } from "typeorm";
 
 import { TagName } from "../../db/entities/tag_names.js";
 import { Tag } from "../../db/entities/tags.js";
-import { TagModel } from "../../graphql/models.js";
 import { QueryResolvers } from "../../graphql/resolvers.js";
+import { TagModel } from "../tag/model.js";
 
 export const searchTags =
   ({ dataSource }: { dataSource: DataSource }): QueryResolvers["searchTags"] =>
