+++ src/resolvers/query/search_videos.ts
@@ -3,8 +3,8 @@ import { DataSource, In, Like } from "typeorm";
 
 import { VideoTitle } from "../../db/entities/video_titles.js";
 import { Video } from "../../db/entities/videos.js";
-import { VideoModel } from "../../graphql/models.js";
 import { QueryResolvers } from "../../graphql/resolvers.js";
+import { VideoModel } from "../video/model.js";
 
 export const searchVideos =
   ({ dataSource }: { dataSource: DataSource }): QueryResolvers["searchVideos"] =>
