+++ src/resolvers/video/index.ts
@@ -7,10 +7,12 @@ import { VideoTag } from "../../db/entities/video_tags.js";
 import { VideoThumbnail } from "../../db/entities/video_thumbnails.js";
 import { VideoTitle as VideoTitleEntity } from "../../db/entities/video_titles.js";
 import { Video } from "../../db/entities/videos.js";
-import { NicovideoVideoSourceModel, TagModel, VideoModel } from "../../graphql/models.js";
 import { Resolvers } from "../../graphql/resolvers.js";
 import { calcVideoSimilarities } from "../../neo4j/video_similarities.js";
 import { addIDPrefix, ObjectType } from "../../utils/id.js";
+import { NicovideoVideoSourceModel } from "../NicovideoVideoSource/model.js";
+import { TagModel } from "../tag/model.js";
+import { VideoModel } from "./model.js";
 
 export const resolveId = ({ id }: VideoModel) => addIDPrefix(ObjectType.Video, id);
 export const resolveHistory = () => ({ nodes: [] });
